requirements.txt
google-genai
python-dotenv
typer